Extract ComponentCommonStyleBuilder from ComponentStyleBuilder
